Module hoprd_inbox::inbox

Defines the common traits and the MessageInbox type.

Structs

MessageInbox

Represents a thread-safe message inbox of messages of type M This type is thin frontend over InboxBackend using 16-bit unsigned integer tags. Each operation also performs purge of the backend.

Traits

InboxBackend

Represents a generic backend trait for the message inbox. Messages M can be tagged or untagged via the type T

Type Aliases

TimestampFn

Represents a simple timestamping function. This is useful if used in WASM or environment which might have different means of measuring time.