hoprd_inbox::ring

Struct RingBufferInboxBackend

pub struct RingBufferInboxBackend<T, M>
where
    T: Copy + Default + PartialEq + Eq + Hash + Send + Sync,
    M: Clone + Send + Sync,
{ /* private fields */ }

Ring buffer based heap-allocated backend.

The capacity must be a power-of-two due to optimizations. Tags T must be represented by a type that’s also a valid key for the HashMap

Implementations

impl<T, M> RingBufferInboxBackend<T, M>
where T: Copy + Default + PartialEq + Eq + Hash + Send + Sync, M: Clone + Send + Sync,

pub fn new(capacity: usize) -> Self

Creates new backend with default timestamping function from std::time.

pub fn count_untagged(&self) -> usize

Counts only the untagged entries.