Type Alias hoprd_inbox::Inbox

pub type Inbox = MessageInbox<RingBufferInboxBackend<Tag, ApplicationData>>;

Aliased Type

struct Inbox { /* private fields */ }

Implementations

impl<B> MessageInbox<B>where B: InboxBackend<Tag, ApplicationData>,

pub fn new(cfg: MessageInboxConfiguration) -> Self

Creates new instance given the configuration.

Uses std::time::SystemTime as timestamping function.

pub fn new_with_time(cfg: MessageInboxConfiguration, time: TimestampFn) -> Self

Creates new instance given the configuration and the timestamping function.

pub async fn push(&self, payload: ApplicationData) -> bool

Push message into the inbox. Returns true if the message has been enqueued, false if it has been excluded based on the configured exluded tags.

pub async fn size(&self, tag: Option<Tag>) -> usize

Number of messages in the inbox with the given tag, or the total number of all messages if no tag is given.

pub async fn pop(&self, tag: Option<Tag>) -> Option<(ApplicationData, Duration)>

Pop the oldest message with the given tag, or the oldest message regardless the tag if it is not given. Returns None if there’s no message with such tag (if given) in the inbox or if the whole inbox is empty (if no tag is given).

pub async fn peek( &self, tag: Option<Tag> ) -> Option<(ApplicationData, Duration)>

Peek the oldest message with the given tag, or the oldest message regardless the tag if it is not given. Returns None if there’s no message with such tag (if given) in the inbox or if the whole inbox is empty (if no tag is given).

pub async fn peek_all( &self, tag: Option<Tag> ) -> Vec<(ApplicationData, Duration)>

Peeks all the messages with the given tag (ordered oldest to latest) or all the messages from the entire inbox (ordered oldest to latest) if no tag is given.

pub async fn pop_all( &self, tag: Option<Tag> ) -> Vec<(ApplicationData, Duration)>

Pops all the messages with the given tag (ordered oldest to latest) or all the messages from the entire inbox (ordered oldest to latest) if no tag is given.